pyproject.toml
[project]
name = "swiftdropship"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-wtf>=1.2.2",
    "flask>=3.1.2",
    "flask-mail>=0.10.0",
    "flask-sqlalchemy>=3.1.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "werkzeug>=3.1.3",
    "numpy>=2.3.3",
    "bleach>=6.2.0",
]
